fixed numeric truncation error in ndpi_main.c:6837 (#1998)

+++ src/include/ndpi_typedefs.h
@@ -1152,7 +1152,7 @@ typedef struct {
 struct ndpi_detection_module_struct {
   NDPI_PROTOCOL_BITMASK detection_bitmask;
 
-  u_int32_t current_ts;
+  u_int64_t current_ts;
   u_int16_t max_packets_to_process;
   u_int16_t num_tls_blocks_to_follow;
   u_int8_t skip_tls_blocks_until_change_cipher:1, enable_ja3_plus:1, _notused:6;
